moved featurization code out of siamese_bert into its own module
