+++ bert/featurization.py
@@ -1,0 +1,66 @@
+from tqdm import tqdm
+
+
+def convert_examples_to_features(examples, max_seq_length, tokenizer):
+    features = []
+
+    for (ex_index, example) in tqdm(enumerate(examples)):
+        feature = convert_single_example(ex_index, example, max_seq_length,
+                                         tokenizer)
+        features.append(feature)
+    return features
+
+
+def convert_single_example(ex_index, example, max_seq_length, tokenizer):
+    """Converts single `InputExample` into a single `InputFeaturesPair`. This
+    differs from `run_classifier.convert_single_example()` in that the two texts
+    are not encoded into the same object"""
+    tokens_a = tokenizer.tokenize(example.text_a)
+    tokens_b = tokenizer.tokenize(example.text_b)
+
+    if len(tokens_a) > max_seq_length - 2:
+        tokens_a = tokens_a[0:(max_seq_length - 2)]
+    if len(tokens_b) > max_seq_length - 2:
+        tokens_b = tokens_b[0:(max_seq_length - 2)]
+
+    tokens_a = ["[CLS]"] + tokens_a + ["[SEP]"]
+    tokens_b = ["[CLS]"] + tokens_b + ["[SEP]"]
+
+    input_ids_a = tokenizer.convert_tokens_to_ids(tokens_a)
+    input_ids_b = tokenizer.convert_tokens_to_ids(tokens_b)
+
+    input_mask_a = [1] * len(input_ids_a)
+    input_mask_b = [1] * len(input_ids_b)
+
+    while len(input_ids_a) < max_seq_length:
+        input_ids_a.append(0)
+        input_mask_a.append(0)
+
+    while len(input_ids_b) < max_seq_length:
+        input_ids_b.append(0)
+        input_mask_b.append(0)
+
+    assert len(input_ids_a) == max_seq_length
+    assert len(input_mask_a) == max_seq_length
+    assert len(input_ids_b) == max_seq_length
+    assert len(input_mask_b) == max_seq_length
+
+    feature = InputFeaturesPair(
+        l_input_ids=input_ids_a,
+        r_input_ids=input_ids_b,
+        l_input_mask=input_mask_a,
+        r_input_mask=input_mask_b)
+    return feature
+
+class InputFeaturesPair(object):
+    """Single pair of features"""
+    def __init__(self,
+                 l_input_ids,
+                 r_input_ids,
+                 l_input_mask,
+                 r_input_mask):
+        self.l_input_ids = l_input_ids
+        self.r_input_ids = r_input_ids
+        self.l_input_mask = l_input_mask
+        self.r_input_mask = r_input_mask
+
